requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.0
Click==7.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.6
django-filter==22.1
django-jazzmin==2.5.0
django-nested-admin==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
future==0.18.2
idna==2.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
pendulum==2.0.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.1
python-decouple==3.6
python-monkey-business==1.0.0
pytz==2022.1
pytzdata==2019.3
redis==4.3.4
requests==2.22.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.13.0
sqlparse==0.4.2
te==1.0.1
uritemplate==4.1.1
urllib3==1.25.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
